完善安装流程，脚本配置分离，规范化脚本配置使用。
* 脚本安装增加prefix设置

* 补充提交，修改machineinstall的bug，删除无用脚本

* 默认安装位置放到配置里，增加ini配置解析脚本

* 统一安装的脚本的存放位置，更改一些错误显示

* 更改更清晰易懂的组件名字，调整更清晰的安装相对路径，为后续兼容配置目录

* 合并publictools和systools

* 备份脚本更改为使用ini配置，暂时去除备份脚本中硬盘检查和发邮件的功能

* 增加configtemplate文件夹的安装

* 工具脚本书写方式整理

* 新增网络检测脚本

* bug修复

* bug修复

* bug修复

* bug修复

* bug

* 去掉敏感信息

* ssh禁止root登录的步骤暂时不使用

* 命令行显示里增加当前时间

* usertools更改配置文件位置到config下

* usertools ssh脚本的配置先不从文件读取

* usertools修复ssh高版本多个-p参数的bug

* usertools修改bashrc的时候删除旧版

* usertools的ssh配置放在config下

* 增加usertools的clean功能

* 修改脚本函数变量为local

* 增加安装日志，更改一个error为warning信息

* 增加log的logrotate配置

* toolsconfig的配置整理一下

* parseini优化效率

* 一个文件夹权限bug

* 整理写法

* 修复mygo函数的一个bug

* sendmail脚本更改

* 邮件脚本整理

* 取消清除去除error

* 系统脚本增加邮件发送

* 更改一个错误为警告

* 更改日志名

* 修改bug

* README更新

+++ source/systools/shutdown_ip.sh
@@ -1,8 +1,14 @@
 #! /usr/bin/bash
 
-host="192.168.5.211"
-usr="root"
-password="19950127"
+if [ $# -lt 3 ];then
+	echo "$0 host usr password"
+	echo "Example: $0 192.168.1.1 test testpass"
+	exit 1
+fi
+
+host=$1
+usr=$2
+password=$3
 cmd="poweroff"
 
 /usr/bin/expect << EOF
